Fix path in file provider for common-auth middleware

+++ init.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 
 echo "[$0] Initializing..."
-cp .env.sample .env
-echo "[$0] Please edit .env file"
+docker network create traefik-network || true
+if [[ ! -f .env ]]; then
+  cp .env.sample .env
+  echo "[$0] Please edit .env file"
+fi
+echo "[$0] Done."
 exit 0
